requirements.txt
# Core ML Libraries
scikit-learn>=1.3.0
xgboost>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Explainability
shap>=0.42.0
lime>=0.2.0.1

# LLM Integration (Optional - for agentic reasoning)
langchain>=0.1.0
langchain-community>=0.0.10
ollama>=0.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
joblib>=1.3.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Web UI
streamlit>=1.28.0

